+++ lib/designs/whatsapp/whatsapp_done_btn.dart
@@ -65,7 +65,6 @@ class _WhatsAppDoneBtnState extends State<WhatsAppDoneBtn> {
     return CupertinoButton(
       onPressed: widget.onPressed,
       padding: const EdgeInsets.symmetric(horizontal: 7),
-      minimumSize: Size.zero,
       child: Text(
         widget.configs.i18n.done,
         style: TextStyle(