+++ lib/features/blur_editor/blur_editor.dart
@@ -13,6 +13,7 @@ import '/core/models/editor_image.dart';
 import '/core/models/init_configs/blur_editor_init_configs.dart';
 import '/core/models/transform_helper.dart';
 import '/core/platform/io/io_helper.dart';
+import '/core/utils/size_utils.dart';
 import '/features/blur_editor/widgets/blur_editor_bottombar.dart';
 import '/shared/controllers/video_controller.dart';
 import '/shared/services/content_recorder/widgets/content_recorder.dart';
@@ -168,6 +169,7 @@ class BlurEditorState extends State<BlurEditor>
 
   /// Represents the selected blur state.
   double get blurFactor => _blurFactor.value;
+
   set blurFactor(double value) {
     _blurFactor.value = value;
   }
@@ -296,9 +298,9 @@ class BlurEditorState extends State<BlurEditor>
                   LayerStack(
                     transformHelper: TransformHelper(
                       mainBodySize:
-                          getMinimumSize(mainBodySize, editorBodySize),
+                          getValidSizeOrDefault(mainBodySize, editorBodySize),
                       mainImageSize:
-                          getMinimumSize(mainImageSize, editorBodySize),
+                          getValidSizeOrDefault(mainImageSize, editorBodySize),
                       transformConfigs: initialTransformConfigs,
                       editorBodySize: editorBodySize,
                     ),
@@ -333,8 +335,10 @@ class BlurEditorState extends State<BlurEditor>
             stream: _uiBlurStream.stream,
             builder: (context, snapshot) {
               return FilteredWidget(
-                width: getMinimumSize(mainImageSize, editorBodySize).width,
-                height: getMinimumSize(mainImageSize, editorBodySize).height,
+                width:
+                    getValidSizeOrDefault(mainImageSize, editorBodySize).width,
+                height:
+                    getValidSizeOrDefault(mainImageSize, editorBodySize).height,
                 configs: configs,
                 image: editorImage,
                 videoPlayer: videoController?.videoPlayer,