+++ lib/shared/widgets/layer/interaction_helper/layer_interaction_helper_widget.dart
@@ -60,6 +60,7 @@ class LayerInteractionHelperWidget extends StatefulWidget
     this.selected = false,
     this.isInteractive = false,
     this.callbacks = const ProImageEditorCallbacks(),
+    this.forceIgnoreGestures = false,
   });
 
   /// The configuration settings for the image editor.
@@ -120,6 +121,13 @@ class LayerInteractionHelperWidget extends StatefulWidget
   /// tooltips.
   final bool isInteractive;
 
+  /// Determines whether gesture interactions should be forcibly ignored.
+  ///
+  /// When set to `true`, all gesture interactions with the associated widget
+  /// will be ignored, regardless of other conditions. This can be useful in
+  /// scenarios where you want to temporarily disable user interaction.
+  final bool forceIgnoreGestures;
+
   /// Indicates whether the layer is selected.
   ///
   /// If true, the layer is highlighted, and interaction buttons are displayed.
@@ -154,6 +162,11 @@ class _LayerInteractionHelperWidgetState
 
   @override
   Widget build(BuildContext context) {
+    if (widget.forceIgnoreGestures) {
+      return IgnorePointer(
+          ignoring: widget.forceIgnoreGestures, child: widget.child);
+    }
+
     String layerId = widget.layerData.id;
     var deferManager = DeferManager.maybeOf(context);